+++ src/main/java/com/parah/mg/frames/operaciones/FrameAportesDetalle.java
@@ -653,11 +653,11 @@ public class FrameAportesDetalle extends JInternalFrame {
 
     private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
         try {
-            int reply = JOptionPane.showConfirmDialog(null, "Realmente desea borrar los registros seleccionados?", title, JOptionPane.YES_NO_OPTION);
+            Integer reply = JOptionPane.showConfirmDialog(null, "Realmente desea borrar los registros seleccionados?", title, JOptionPane.YES_NO_OPTION);
             if (reply == JOptionPane.YES_OPTION) {
                 int[] selected = masterTable.getSelectedRows();
-                for (int idx = 0; idx < selected.length; idx++) {
-                    int index = masterTable.convertRowIndexToModel(selected[idx]);
+                for (Integer idx = 0; idx < selected.length; idx++) {
+                    Integer index = masterTable.convertRowIndexToModel(selected[idx]);
                     TblEventoDetalle t = listEventoDetalle.get(index);
                     entityManager.remove(t);
                 }
@@ -688,7 +688,7 @@ public class FrameAportesDetalle extends JInternalFrame {
 
             entityManager.persist(t);
             listEventoDetalle.add(t);
-            int row = listEventoDetalle.size() - 1;
+            Integer row = listEventoDetalle.size() - 1;
             masterTable.setRowSelectionInterval(row, row);
             masterTable.scrollRectToVisible(masterTable.getCellRect(row, 0, true));
             cboForma.setSelectedIndex(0);
@@ -721,7 +721,7 @@ public class FrameAportesDetalle extends JInternalFrame {
              if ((Long) entityManager.createQuery("SELECT COUNT(t) FROM TblTransferencias t WHERE t.cobrado = true AND t.idEvento.id = " + String.valueOf(currentEvento.getId())).getSingleResult() > 0L) {
              JOptionPane.showMessageDialog(null, "Los cambios realizados a este evento no tendrán efecto sobre los pagos realizados porque ya hay transferencias cobradas para este evento. Deberá modificar las transferencias y los recibos manualmente.");
              } else {
-             int res = entityManager.createQuery("DELETE FROM TblTransferencias t WHERE t.idEvento.id = " + String.valueOf(currentEvento.getId())).executeUpdate();
+             Integer res = entityManager.createQuery("DELETE FROM TblTransferencias t WHERE t.idEvento.id = " + String.valueOf(currentEvento.getId())).executeUpdate();
              res = entityManager.createQuery("DELETE FROM TblRecibos t WHERE t.idEvento.id = " + String.valueOf(currentEvento.getId())).executeUpdate();
 
              listEventoDetalle.stream().forEach((t) -> {
@@ -772,7 +772,7 @@ public class FrameAportesDetalle extends JInternalFrame {
                             indexAsientoAporte = 1;
                             indexAsientoDonacion = 0;
                         }
-                        ((List<TblAsientos>) evd.getTblAsientosCollection()).get(indexAsientoAporte).setMonto(evd.getMonto() * evd.getIdEvento().getPorcentajeAporte() / 100);
+                        ((List<TblAsientos>) evd.getTblAsientosCollection()).get(indexAsientoAporte).setMonto(((Long) (evd.getMonto().longValue() * evd.getIdEvento().getPorcentajeAporte().longValue() / 100)).intValue());
                         ((List<TblAsientos>) evd.getTblAsientosCollection()).get(indexAsientoDonacion).setMonto(evd.getMonto() - ((List<TblAsientos>) evd.getTblAsientosCollection()).get(indexAsientoAporte).getMonto());
                         entityManager.merge(evd);
                     } else if (evd.getTblAsientosCollection().isEmpty()) {
@@ -787,7 +787,7 @@ public class FrameAportesDetalle extends JInternalFrame {
                         asientoAporte.setIdCentroDeCosto(evd.getIdEvento().getIdCentroDeCosto());
                         asientoAporte.setIdCuentaContableDebe(cuentasContablesPorDefecto.getIdCuentaACobrar());
                         asientoAporte.setIdCuentaContableHaber(cuentasContablesPorDefecto.getIdCuentaAportes());
-                        asientoAporte.setMonto(evd.getMonto() * evd.getIdEvento().getPorcentajeAporte() / 100);
+                        asientoAporte.setMonto(((Long) (evd.getMonto().longValue() * evd.getIdEvento().getPorcentajeAporte().longValue() / 100)).intValue());
                         asientoAporte.setIdUser(currentUser.getUser());
 
                         ts.add(asientoAporte);
@@ -853,7 +853,7 @@ public class FrameAportesDetalle extends JInternalFrame {
                     cboEntidad.setSelectedItem(valueList.get(0));
                     txtCtaCte.setBackground(Color.green);
                     txtCtaCte.requestFocus();
-                    int end = txtCtaCte.getSelectionEnd();
+                    Integer end = txtCtaCte.getSelectionEnd();
                     txtCtaCte.setSelectionStart(end);
                     txtCtaCte.setSelectionEnd(end);
                 }
@@ -977,7 +977,7 @@ public class FrameAportesDetalle extends JInternalFrame {
                         t.setMonto(miembro.getAporteMensual());
                         t.setIdUser(currentUser.getUser());
                         listEventoDetalle.add(t);
-                        int row = listEventoDetalle.size() - 1;
+                        Integer row = listEventoDetalle.size() - 1;
                         masterTable.setRowSelectionInterval(row, row);
                         //masterTable.scrollRectToVisible(masterTable.getCellRect(row, 0, true));
                     }
@@ -1008,7 +1008,7 @@ public class FrameAportesDetalle extends JInternalFrame {
                     return;
                 }
                 save();
-                int reply = JOptionPane.showConfirmDialog(null, "Desea crear un nuevo registro?", title, JOptionPane.YES_NO_OPTION);
+                Integer reply = JOptionPane.showConfirmDialog(null, "Desea crear un nuevo registro?", title, JOptionPane.YES_NO_OPTION);
                 if (reply == JOptionPane.YES_OPTION) {
                     newDetalle();
                 }
